refactor: :art: Refactored update method of NegotiationDAO class
