+++ backend/src/db/ComunicationDB/NegotiaionDAO/index.js
@@ -1,7 +1,7 @@
 const { InvalidInputError, NotFoundError } = require("../../../CustomErrors")
 const { checkInvalidInputsErrors, checkForeignKeyError } = require("../../../utils")
 const { dbGet, dbAll } = require("../../utils/dbutils")
-const { validateFilters, createSqlFilterForSelect } = require("./utils")
+const { validateFilters, createSqlFilterForSelect, createSqlAndArgsForUpdate } = require("./utils")
 
 /**
  * Class responsible for comunication with DB to manipulate the negotiations table, that stores
@@ -101,52 +101,31 @@ class NegotiationDAO {
 
     //method for updating an specific negotiation
     static async update(negotiationToBeUpdated) {
+        //check if argument has a negotiation id
         const { id } = negotiationToBeUpdated
-        //check if it has id
         if(!id) throw new InvalidInputError('The update method requires the id of the negotiotion to be updated', ['negotiationToBeUpdated.id'])
         
-        const propToColMap = {
-            id: 'id', 
-            userId: 'user_id', 
-            stockTicker: 'stock_ticker', 
-            negotiatedQty: 'negotiated_qty', 
-            negotiatedPrice: 'negotiated_price', 
-            negotiationType: 'negotiation_type', 
-            negotiationDate: 'negotiation_date' 
-        }
-
+        //Transform entered object in array for manipulation
         const negotiationEntries = Object.entries(negotiationToBeUpdated)
+
+        //Removing negotiation id so items in entries array are all columns to be updated
         const idIndex = negotiationEntries.findIndex(entry => entry[0] === 'id')
         negotiationEntries.splice(idIndex, 1)
-        
-        //check if it has at least one column besides id
+
+        //check if it has at least one column besides id by checking if there are any entries left after the removal of id
         if(negotiationEntries.length <= 0) throw new InvalidInputError('Enter at least one field of the negotiation to be updated', ['negotiationToBeUpdated'])
-
-        //build sql and list of aguments
-        const baseSQL = `UPDATE negotiations SET `
-        const filterSQL = `WHERE id=?`
-        const returnValueSQL = 'RETURNING id'
-        const listOfSqlUpdates = []
-        const sqlQueryArgumentList = []
-        negotiationEntries.forEach(entry => {
-            const [propToUpdate, updatedValue] = entry
-            const columnToUpdate = propToColMap[propToUpdate]
-            //throw invalid input error if negotiation object arg has property that is invalid
-            if(!columnToUpdate) throw new InvalidInputError('One of the properties to update is invalid', [propToUpdate])
-            listOfSqlUpdates.push(`${columnToUpdate}=?`)
-            sqlQueryArgumentList.push(updatedValue)
-        })
-
-        const updateSQL = listOfSqlUpdates.join(', ')
-
-        const sql = baseSQL + updateSQL + filterSQL + returnValueSQL
+        
+        //create sql string and arguments list to run query using util function
+        const { sql, sqlQueryArgumentList } = createSqlAndArgsForUpdate(negotiationEntries)
+        
+        //add id to the end of the argument list so it enters as the filter in the sql query
         sqlQueryArgumentList.push(id)
 
         try {
             //run sql
             const updatedId = await dbGet(sql, sqlQueryArgumentList)
     
-            //throw not found error if id not in db
+            //throw not found error if negotiation id is not in db
             if(!updatedId) throw new NotFoundError('Negotiation not found in our database. Please check the entered id')
             
         } catch (error) {
@@ -156,7 +135,6 @@ class NegotiationDAO {
             //throw error so it can be caught at controller layer
             throw error
         }
-
 
     }
